Rename fields in CiliumFlowLogging
